requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu114
cog
torch
torchvision
torchaudio==2.3.1
diffusers==0.32.2
easydict==1.13
einops==0.8.1
peft==0.14.0
pillow==11.0.0
protobuf==5.29.3
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
transformers==4.49.0
googleapis-common-protos
google-generativeai
huggingface_hub
google-api-core
accelerate